handle more ng exit codes
comments from #87:

* cbed5408dddccae5589bcef73c5aeedeed06b8e5
For some reasons Nailgun thinks that it's server
broke the connection and returns 227 instead of 0
We'll just handle this here instead of fixing
the nailgun client itself.

* 7810d4a44563f7a5dddff7fca2e11e7191979bbe
It also sometimes breaks on 229 on the devbox.
To complete this whole madness and made future
'fixes' easier I added this nice little array...